refactor: move daemon host primitives to owner package
