+++ packages/daemon/src/lifecycle/daemon-log-file-store.ts
