+++ packages/daemon/src/runtime/canonical-harness-root.ts
@@ -1,3 +1,4 @@
+// @slice-activation PLT-Boundary W2 exports the daemon-owned canonical harness root to CLI consumers.
 import { existsSync, readFileSync, realpathSync, statSync } from "node:fs";
 import path from "node:path";
 import { resolveHarnessLayout, type HarnessLayoutInput } from "@harness-anything/kernel";
